feat(CEPA-179): Add Topic course - Create add course topic - Add step add course topic for create course - Set status for video mapper - Add price for course detail

+++ src/main/java/com/example/courseservice/data/entities/Rating.java
@@ -26,9 +26,9 @@ public class Rating {
     @Id
     @SequenceGenerator(name = "rating_sequence", sequenceName = "rating_sequence", allocationSize = 1)
     @GeneratedValue(strategy = GenerationType.AUTO, generator = "rating_sequence")
-    private long id;
+    private Long id;
 
-    private int rate;
+    private Float rate;
 
     private LocalDateTime createDate;
